Fix combination of --create, --append, --directory
In copy-out mode, when appending, change_dir() would be called twice,
because process_copy_in() calls change_dir().  If --directory=DIR was
used with a relative path, this would result in incorrectly trying to
change to DIR/DIR.  To fix this, change_dir() is only called
explicitly when not appending.

Patch by Alyssa Ross <hi@alyssa.is>

* src/copyout.c (process_copy_out): Don't call change_dir if
append_flag is set: it is going to be called by process_copy_in.