Build a "literate" executable.
Literate executable incorporates the source code and is able to
reproduce it on demand. See https://queue.acm.org/detail.cfm?id=3570938.
Instead of the approach proposed in the article, this build first
creates a plain executable (stage1), which is then used to produce
a cpio archive of the source in the form suitable for initializing
a char array. Finally, the created file is used to build a literate
executable.

The --dump-source option instructs the executable to dump the source
cpio archive on the standard output. Optional argument gives the
name of the file to save the archive in.